[ci skip] [WIP] Rebased and fixing builds

+++ src/jet/point_parallel_hash_grid_searcher3.cpp
@@ -372,7 +372,7 @@ void PointParallelHashGridSearcher3::deserialize(
 
     // Copy simple data
     auto res = fbsToJet(*fbsSearcher->resolution());
-    _resolution = {res.x, res.y, res.z};
+    _resolution = {(ssize_t)res.x, (ssize_t)res.y, (ssize_t)res.z};
     _gridSpacing = fbsSearcher->gridSpacing();
 
     // Copy points
